requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.2
Pillow>=10.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.3
email-validator==2.1.1
Flask-Babel==4.0.0